Make the code compile for the win32

+++ include/connection_id.h
@@ -2,7 +2,8 @@
 // Created by mswiercz on 27.11.2021.
 //
 #pragma once
+#include <cstdint>
 
 namespace auction_house {
-using ConnectionId = int; // socket desc
+using ConnectionId = int64_t; // this type can store unix and winsock socket fd
 }
